+++ lib/tv_textfield.dart
@@ -1,5 +1,5 @@
 export 'src/tv_text_field.dart';
-export 'src/tv_text_field_scope.dart';
+export 'src/tv_text_field_scope.dart' show TvTextFieldScope, initializeTvTextField;
 export 'src/platform/flutter_tv_text_field.dart'
     show resolveTvTextFieldImplementation;
 export 'tv_textfield_stub.dart' show TvTextfieldPluginStub;